% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regSmooth1.1.R
\name{divided.diff}
\alias{divided.diff}
\title{Create divided difference matrix}
\usage{
divided.diff(x, d)
}
\arguments{
\item{x}{vector to process}

\item{d}{Dimension}
}
\value{
Output is divided difference matrix for use internally in regSmooth
}
\description{
Create divided difference matrix
}
\keyword{internal}

