% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regSmooth1.1.R
\name{regSmoothAuto}
\alias{regSmoothAuto}
\title{Data smoothing by regularisation.}
\usage{
regSmoothAuto(x, y, lambda, d = 2, relative = F, midpr = F,
  maxit = 1000, guess.lambda = 0.001, ...)
}
\arguments{
\item{x}{x values of data to smooth. Numeric vector.If xhat is not supplied, then x must be strictly monotonic increasing.}

\item{y}{y values of data to smooth. Numeric vector. Must be same length as x.}

\item{lambda}{smoothing parameter (positive real numeric). Higher values give greater smoothing. Smaller values follow data more closely.}

\item{d}{Optional order of smoothing derivative. Default=2 (2nd order derivatives).}

\item{relative}{Logical (default = False) flag to use relative differences for the goodness of fit term.  Conflicts with the weights (w) optional argument. If relative=True and weights are provided as vector w, then a warning will be issued, and relative set to False.}

\item{midpr}{Logical (default = False) flag to use the midpoint rule for the integration terms rather than a direct sum. This option conflicts with the option "xhat". If midpr=True and xhat is provided, then a warning will be issued, and midpr set to False.}

\item{maxit}{Maximum number of iterations to use when solving for lambda.}

\item{guess.lambda}{Starting value for lambda optimisation.}

\item{...}{Additional optional parameters to pass to regSmooth:
w =  Optional weighting values for fitting each point in the data (numeric vector). If not provided, then all data points are given equal weighting. Must have same length as x and y.
xhat = An optional numeric vector of x-values to use for the smooth curve. If provided, xhat must be strictly monotonically increasing, and must at least span the data. If xhat is not provided, then the smoothed y values will be have x coordinates at x.}
}
\value{
Provides a list of two elements:
 1) yhat = A numeric vector (yhat) of smoothed y values. Vector is of same length as xhat (if provided), or x if not. If xhat is provided, then the smoothed function is given by the data points (xhat,yhat), otherwise, the smoothed curve is (x,yhat).
 2) variance = the variance of the fitted (smoothed model).
}
\description{
Function to smooth data using Tikhonov regularisation,
with optional optimisation of the regularisation parameter by generalised cross validation.
}
\details{
This function is based on the method of Eilers (2003), and its implementation by Stickel (2010).
}
\examples{
  x <- 1:1000
  y <- sin(x/100) + 0.3*cos(x/10)              # Generate a curve
  y.noisy <- y + rnorm(1000, sd=0.1)           # Add some noise
  y.smoothed <- regSmoothAuto(x=x, y=y.noisy)  # Smooth the noisy data
  plot  (y.noisy ~ x)                          # Plot the noisy data
  lines (y.smoothed$yhat ~ x, col="red", lw=2) # Draw the smoothed curve in red
  lines (y ~ x, col="darkgreen", lw=2, lt=1)   # Original function (without noise) in green
}
\author{
Dominic Woolf.
  d.woolf@cornell.edu
}
\references{
Eilers, P. (2003). A Perfect Smoother. Anal. chem. 75, 3631-3636.

  Stickel, J. (2010). Data smoothing and numerical differentiation by a regularization method. Comp. chem. eng. 34.4 467-475.
}
\seealso{
\code{\link{regSmooth}}
}
\keyword{Smoothing}

